+++ server/src/main/java/com/ones/admin/hr/mapper/HrEmployeeContractMapper.java
@@ -1,0 +1,7 @@
+package com.ones.admin.hr.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.ones.admin.hr.entity.HrEmployeeContractEntity;
+
+public interface HrEmployeeContractMapper extends BaseMapper<HrEmployeeContractEntity> {
+}